+++ packages/admin-stories/src/admin/form/Radio.tsx
@@ -1,4 +1,4 @@
-import { Field, FieldContainer, FinalFormInput, FinalFormRadio } from "@comet/admin";
+import { Field, FieldContainer, FinalFormRadio } from "@comet/admin";
 import { FormControlLabel } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
@@ -11,9 +11,8 @@ function Story() {
                 onSubmit={(values) => {
                     //
                 }}
-                render={({ handleSubmit, pristine, invalid }) => (
+                render={({ handleSubmit }) => (
                     <form onSubmit={handleSubmit}>
-                        <Field name="name" label="Name" type="text" component={FinalFormInput} fullWidth />
                         <FieldContainer label="Foo">
                             <Field name="foo" type="radio" value="foo1">
                                 {(props) => <FormControlLabel label="Foo1" control={<FinalFormRadio {...props} />} />}